Move desktop command payload types into api/types/desktop.ts
Keep protocol.ts to the WS envelope; the desktop/check_update and desktop/update
response shapes belong beside the other feature payload type files.
